chainlit_image/foundational-llm-chat_app/pyproject.toml
[project]
name = "foundational-llm-chat-app"
version = "0.1.0"
description = "Example Chainlit-based UI for talking to Foundation Models on Amazon Bedrock"
requires-python = ">=3.13"
dependencies = [
    "boto3==1.38.22",
    "chainlit==2.5.5",
    "pillow==11.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]
